+++ server/src/payment/controllers/stripe.controller.ts
@@ -5,39 +5,14 @@ import {
   Req,
   RawBodyRequest,
   Get,
-  Query,
   Param,
-  Patch,
 } from '@nestjs/common';
 import { stripe } from 'src/clients';
 import { Request } from 'express';
-import { delay } from 'src/common/helpers';
 import { UserService } from 'src/user/services/user.service';
 
 interface IBody {
   line_items: [{ quantity: number; price: string }];
-}
-
-interface ICreatePaymentIntentBody extends IBody {
-  customerId?: string;
-  userData?: {
-    name: string;
-    email: string;
-    country: string;
-    postalCode: string;
-  };
-}
-
-async function calculateTotalAmount(line_items) {
-  let total = 0;
-
-  for (const item of line_items) {
-    const price = await stripe.prices.retrieve(item.price);
-    const unitAmount = price.unit_amount; // cents/pennies
-    total += unitAmount * item.quantity;
-  }
-
-  return total;
 }
 
 @Controller()
@@ -91,39 +66,6 @@ export class CreateStripeCheckoutSessionController {
     }
   }
 
-  @Post('stripe/payment-intents')
-  async createPaymentIntent(@Body() body: ICreatePaymentIntentBody) {
-    const stripeCustomerId = body.customerId
-      ? (await this.userService.findOne(body.customerId)).stripeCustomerId
-      : null;
-
-    const paymentIntent = await stripe.paymentIntents.create({
-      amount: await calculateTotalAmount(body.line_items),
-      currency: 'usd',
-      automatic_payment_methods: { enabled: true },
-      ...(body.customerId && {
-        customer: stripeCustomerId,
-      }),
-      ...(body.userData && {
-        receipt_email: body.userData.email,
-        shipping: {
-          name: body.userData.name,
-          address: {
-            country: body.userData.country,
-            postal_code: body.userData.postalCode,
-          },
-        },
-      }),
-    });
-
-    await delay(5000);
-
-    return {
-      clientSecret: paymentIntent.client_secret,
-      paymentIntentId: paymentIntent.id,
-    };
-  }
-
   @Post('stripe/setup-intents')
   async createSetupIntent(@Body() body) {
     const user = await this.userService.findOne(body.userId);
@@ -146,18 +88,5 @@ export class CreateStripeCheckoutSessionController {
     });
 
     return paymentMethods.data;
-  }
-
-  @Patch('stripe/payment-intents/:paymentIntentId')
-  async updatePaymentIntent(
-    @Param('paymentIntentId') paymentIntentId: string,
-    @Body() body,
-  ) {
-    const paymentIntent = await stripe.paymentIntents.update(
-      paymentIntentId,
-      body,
-    );
-
-    return paymentIntent.client_secret;
   }
 }
